support building wheel from sdist

+++ hatch_build.py
@@ -2,13 +2,14 @@
 
 import shutil
 import tarfile
+from contextlib import contextmanager
 from pathlib import Path
-from typing import Any
+from typing import Any, Callable
 
 from hatchling.builders.hooks.plugin.interface import BuildHookInterface
 
 
-def common_root(names: list[str]):
+def common_root(names: list[str], abort: Callable[[str], None]) -> str:
     common = None
     for name in names:
         if common is None:
@@ -20,25 +21,37 @@ def common_root(names: list[str]):
     return common
 
 
+@contextmanager
+def unpack_definitions(abort: Callable[[str], None]) -> str:
+    if (root := Path("eccodes-cosmo-resources")).exists():
+        yield str(root / "definitions")
+        return
+
+    [path] = Path(".").glob("eccodes_definitions*.tar.bz2")
+    with tarfile.open(path, "r:bz2") as tar:
+        root = common_root(tar.getnames(), abort)
+        if not root.startswith("definitions"):
+            abort("")
+        tar.extractall(filter="data")
+    try:
+        yield root
+    finally:
+        shutil.rmtree(root)
+
+
 class CustomBuildHook(BuildHookInterface):
     def initialize(
         self,
         version: str,
         build_data: dict[str, Any],
     ) -> None:
-        [path] = Path(".").glob("eccodes_definitions*.tar.bz2")
-        with tarfile.open(path, "r:bz2") as tar:
-            root = common_root(tar.getnames())
-            if not root.startswith("definitions"):
-                raise RuntimeError
-            tar.extractall(filter="data")
-        shutil.copytree(
-            root,
-            "tmp",
-            symlinks=False,
-            dirs_exist_ok=True,
-        )
-        shutil.rmtree(root)
+        with unpack_definitions(self.app.abort) as root:
+            shutil.copytree(
+                root,
+                "tmp",
+                symlinks=False,
+                dirs_exist_ok=True,
+            )
 
     def finalize(
         self,
